Completed refactor to individual files, all tests passing
