+++ src/dbc.rs
@@ -8,42 +8,50 @@ use std::convert::TryFrom;
 
 use derive_getters::Getters;
 
+use crate::attributes::*;
+use crate::env_variables::*;
+use crate::message::*;
+use crate::nodes::*;
+use crate::signal::*;
 use crate::DBCString;
 use crate::Error;
-use crate::message::*;
-use crate::signal::*;
-use crate::nodes::*;
-use crate::env_variables::*;
-use crate::attributes::*;
 
 use crate::parser;
 
 use nom::{
     branch::{alt, permutation},
-    bytes::complete::{tag, take_till, take_while, take_while1},
-    character::complete::{self, char, line_ending, multispace0, space0, space1},
-    combinator::{map, opt, value},
-    error::{ErrorKind, ParseError},
-    multi::{many0, many_till, separated_list0},
-    number::complete::double,
+    bytes::complete::tag,
+    character::complete::{self, line_ending, multispace0, space0, space1},
+    combinator::{map, opt},
+    multi::{many0, separated_list0},
     sequence::preceded,
-    AsChar, IResult, InputTakeAtPosition,
+    IResult,
 };
 
 /// Baudrate of network in kbit/s
 #[derive(Copy, Clone, Debug, PartialEq)]
 #[cfg_attr(feature = "serde", derive(Serialize))]
-pub struct Baudrate(pub (crate) u64);
+pub struct Baudrate(pub(crate) u64);
 
 impl DBCString for Baudrate {
     fn dbc_string(&self) -> String {
-        return self.0.to_string()
+        return self.0.to_string();
     }
 
     fn parse(s: &str) -> nom::IResult<&str, Self> {
-        let bd = Baudrate(500000);
-        return IResult::Ok(("rest", bd))
+        let (s, baudrate) = map(complete::u64, Baudrate)(s)?;
+        Ok((s, baudrate))
     }
+}
+
+fn bit_timing(s: &str) -> IResult<&str, Vec<Baudrate>> {
+    let (s, _) = multispace0(s)?;
+    let (s, _) = tag("BS_:")(s)?;
+    let (s, baudrates) = opt(preceded(
+        parser::ms1,
+        separated_list0(parser::comma, map(complete::u64, Baudrate)),
+    ))(s)?;
+    Ok((s, baudrates.unwrap_or_default()))
 }
 
 /// Version generated by DB editor
@@ -53,12 +61,13 @@ pub struct Version(pub String);
 
 impl DBCString for Version {
     fn dbc_string(&self) -> String {
-        return format!("VERSION {}", self.0)
+        return format!("VERSION {}", self.0);
     }
 
     fn parse(s: &str) -> IResult<&str, Self>
-        where
-            Self: Sized {
+    where
+        Self: Sized,
+    {
         let (s, _) = multispace0(s)?;
         let (s, _) = tag("VERSION")(s)?;
         let (s, _) = parser::ms1(s)?;
@@ -68,18 +77,28 @@ impl DBCString for Version {
     }
 }
 
+#[test]
+fn version_test() {
+    let def = "VERSION \"HNPBNNNYNNNNNNNNNNNNNNNNNNNNNNNNYNYYYYYYYY>4>%%%/4>'%**4YYY///\"\n";
+    let version_exp =
+        Version("HNPBNNNYNNNNNNNNNNNNNNNNNNNNNNNNYNYYYYYYYY>4>%%%/4>'%**4YYY///".to_string());
+    let (_, version) = Version::parse(def).unwrap();
+    assert_eq!(version_exp, version);
+}
+
 #[derive(Clone, Debug, PartialEq)]
 #[cfg_attr(feature = "serde", derive(Serialize))]
 pub struct Symbol(pub String);
 
 impl DBCString for Symbol {
     fn dbc_string(&self) -> String {
-        return self.0.to_string()
+        return self.0.to_string();
     }
 
     fn parse(s: &str) -> IResult<&str, Self>
-        where
-            Self: Sized {
+    where
+        Self: Sized,
+    {
         let (s, _) = space1(s)?;
         let (s, symbol) = parser::c_ident(s)?;
         let (s, _) = line_ending(s)?;
@@ -87,54 +106,80 @@ impl DBCString for Symbol {
     }
 }
 
+#[test]
+fn new_symbols_test() {
+    let def = "NS_ :
+            NS_DESC_
+            CM_
+            BA_DEF_
+
+        ";
+    let symbols_exp = vec![
+        Symbol("NS_DESC_".to_string()),
+        Symbol("CM_".to_string()),
+        Symbol("BA_DEF_".to_string()),
+    ];
+    let (_, symbols) = new_symbols(def).unwrap();
+    assert_eq!(symbols_exp, symbols);
+}
+
+pub(crate) fn new_symbols(s: &str) -> IResult<&str, Vec<Symbol>> {
+    let (s, _) = multispace0(s)?;
+    let (s, _) = tag("NS_ :")(s)?;
+    let (s, _) = space0(s)?;
+    let (s, _) = line_ending(s)?;
+    let (s, symbols) = many0(Symbol::parse)(s)?;
+    Ok((s, symbols))
+}
+
 #[derive(Clone, Debug, PartialEq, Getters)]
 #[cfg_attr(feature = "serde", derive(Serialize))]
 pub struct DBC {
     /// Version generated by DB editor
-    pub (crate) version: Version,
-    pub (crate) new_symbols: Vec<Symbol>,
+    pub(crate) version: Version,
+    pub(crate) new_symbols: Vec<Symbol>,
     /// Baud rate of network
-    pub (crate) bit_timing: Option<Vec<Baudrate>>,
+    pub(crate) bit_timing: Option<Vec<Baudrate>>,
     /// CAN network nodes
-    pub (crate) nodes: Vec<Node>,
+    pub(crate) nodes: Vec<Node>,
     /// Global value table
-    pub (crate) value_tables: Vec<ValueTable>,
+    pub(crate) value_tables: Vec<ValueTable>,
     /// CAN message (frame) details including signal details
-    pub (crate) messages: Vec<Message>,
-    pub (crate) message_transmitters: Vec<MessageTransmitter>,
-    pub (crate) environment_variables: Vec<EnvironmentVariable>,
-    pub (crate) environment_variable_data: Vec<EnvironmentVariableData>,
-    pub (crate) signal_types: Vec<SignalType>,
+    pub(crate) messages: Vec<Message>,
+    pub(crate) message_transmitters: Vec<MessageTransmitter>,
+    pub(crate) environment_variables: Vec<EnvironmentVariable>,
+    pub(crate) environment_variable_data: Vec<EnvironmentVariableData>,
+    pub(crate) signal_types: Vec<SignalType>,
     /// Object comments
-    pub (crate) comments: Vec<Comment>,
-    pub (crate) attribute_definitions: Vec<AttributeDefinition>,
+    pub(crate) comments: Vec<Comment>,
+    pub(crate) attribute_definitions: Vec<AttributeDefinition>,
     // undefined
     // sigtype_attr_list: SigtypeAttrList,
-    pub (crate) attribute_defaults: Vec<AttributeDefault>,
-    pub (crate) attribute_values: Vec<AttributeValueForObject>,
+    pub(crate) attribute_defaults: Vec<AttributeDefault>,
+    pub(crate) attribute_values: Vec<AttributeValueForObject>,
     /// Encoding for signal raw values
-    pub (crate) value_descriptions: Vec<ValueDescription>,
+    pub(crate) value_descriptions: Vec<ValueDescription>,
     // obsolete + undefined
     // category_definitions: Vec<CategoryDefinition>,
     // obsolete + undefined
     //categories: Vec<Category>,
     // obsolete + undefined
     //filter: Vec<Filter>,
-    pub (crate) signal_type_refs: Vec<SignalTypeRef>,
+    pub(crate) signal_type_refs: Vec<SignalTypeRef>,
     /// Signal groups define a group of signals within a message
-    pub (crate) signal_groups: Vec<SignalGroups>,
-    pub (crate) signal_extended_value_type_list: Vec<SignalExtendedValueTypeList>,
+    pub(crate) signal_groups: Vec<SignalGroups>,
+    pub(crate) signal_extended_value_type_list: Vec<SignalExtendedValueTypeList>,
     /// Extended multiplex attributes
-    pub (crate) extended_multiplex: Vec<ExtendedMultiplex>,
+    pub(crate) extended_multiplex: Vec<ExtendedMultiplex>,
 }
 
 #[allow(dead_code)]
 fn dbc_vec_to_string<T: DBCString>(dbc_objects: &Vec<T>, delimiter: &str) -> String {
     dbc_objects
-    .into_iter()
-    .map(|sym| sym.dbc_string())
-    .collect::<Vec<String>>()
-    .join(delimiter)
+        .into_iter()
+        .map(|sym| sym.dbc_string())
+        .collect::<Vec<String>>()
+        .join(delimiter)
 }
 
 impl DBCString for DBC {
@@ -146,7 +191,6 @@ impl DBCString for DBC {
         // Symbols
         file_str += "NS_ :\n";
         file_str += &dbc_vec_to_string::<Symbol>(&self.new_symbols, "\n    ");
-        
 
         // Baudrates
         match &self.bit_timing {
@@ -166,10 +210,11 @@ impl DBCString for DBC {
         // Messages
         file_str += &dbc_vec_to_string::<Message>(&self.messages, "\n");
         file_str += &dbc_vec_to_string::<MessageTransmitter>(&self.message_transmitters, " ");
-        
+
         // Environment Variables
         file_str += &dbc_vec_to_string::<EnvironmentVariable>(&self.environment_variables, "\n");
-        file_str += &dbc_vec_to_string::<EnvironmentVariableData>(&self.environment_variable_data, "\n");
+        file_str +=
+            &dbc_vec_to_string::<EnvironmentVariableData>(&self.environment_variable_data, "\n");
 
         // Signal Types
         file_str += &dbc_vec_to_string::<SignalType>(&self.signal_types, " ");
@@ -188,17 +233,21 @@ impl DBCString for DBC {
         // Signal Attributes
         file_str += &dbc_vec_to_string::<SignalTypeRef>(&self.signal_type_refs, "\n");
         file_str += &dbc_vec_to_string::<SignalGroups>(&self.signal_groups, "\n");
-        file_str += &dbc_vec_to_string::<SignalExtendedValueTypeList>(&self.signal_extended_value_type_list, "\n");
+        file_str += &dbc_vec_to_string::<SignalExtendedValueTypeList>(
+            &self.signal_extended_value_type_list,
+            "\n",
+        );
 
         // Multiplex
         file_str += &dbc_vec_to_string::<ExtendedMultiplex>(&self.extended_multiplex, "\n");
 
-        return file_str
+        return file_str;
     }
 
     fn parse(s: &str) -> IResult<&str, Self>
-        where
-            Self: Sized {
+    where
+        Self: Sized,
+    {
         let (
             s,
             (
@@ -237,11 +286,11 @@ impl DBCString for DBC {
             many0(AttributeDefinition::parse),
             many0(AttributeDefault::parse),
             many0(AttributeValueForObject::parse),
-            many0(value_descriptions),
-            many0(signal_type_ref),
-            many0(signal_groups),
-            many0(signal_extended_value_type_list),
-            many0(extended_multiplex),
+            many0(ValueDescription::parse),
+            many0(SignalTypeRef::parse),
+            many0(SignalGroups::parse),
+            many0(SignalExtendedValueTypeList::parse),
+            many0(ExtendedMultiplex::parse),
         ))(s)?;
         let (s, _) = multispace0(s)?;
         Ok((
@@ -283,7 +332,7 @@ impl DBC {
     #[deprecated(since = "4.0.0", note = "please use `DBC::try_from` instead")]
     #[allow(clippy::result_large_err)]
     pub fn from_str(dbc_in: &str) -> Result<DBC, Error> {
-        let (remaining, dbc) = parser::dbc(dbc_in).map_err(Error::Nom)?;
+        let (remaining, dbc) = DBC::parse(dbc_in).map_err(Error::Nom)?;
         if !remaining.is_empty() {
             return Err(Error::Incomplete(dbc, remaining));
         }
@@ -429,7 +478,7 @@ impl<'a> TryFrom<&'a str> for DBC {
     type Error = Error<'a>;
 
     fn try_from(dbc_in: &'a str) -> Result<Self, Self::Error> {
-        let (remaining, dbc) = parser::dbc(dbc_in).map_err(Error::Nom)?;
+        let (remaining, dbc) = DBC::parse(dbc_in).map_err(Error::Nom)?;
         if !remaining.is_empty() {
             return Err(Error::Incomplete(dbc, remaining));
         }
@@ -544,38 +593,41 @@ impl Comment {
 impl DBCString for Comment {
     fn dbc_string(&self) -> String {
         return match self {
-            Self::Node{node_name, comment} => {
-                format!("BU_ {} \"{}\"",
-                    node_name,
-                    comment
-                )
-            },
-            Self::Message{message_id, comment} => {
-                format!("BO_ {} \"{}\"",
-                    message_id.dbc_string(),
-                    comment,
-                )
-            },
-            Self::Signal{message_id, signal_name, comment} => {
-                format!("SG_ {} {} \"{}\"",
+            Self::Node { node_name, comment } => {
+                format!("BU_ {} \"{}\"", node_name, comment)
+            }
+            Self::Message {
+                message_id,
+                comment,
+            } => {
+                format!("BO_ {} \"{}\"", message_id.dbc_string(), comment,)
+            }
+            Self::Signal {
+                message_id,
+                signal_name,
+                comment,
+            } => {
+                format!(
+                    "SG_ {} {} \"{}\"",
                     message_id.dbc_string(),
                     signal_name,
                     comment
                 )
-            },
-            Self::EnvVar{env_var_name, comment} => {
-                format!("EV_ {} \"{}\"",
-                    env_var_name,
-                    comment,
-                )
-            },
-            Self::Plain{comment} => format!("\"{}\"", comment),
-        }
+            }
+            Self::EnvVar {
+                env_var_name,
+                comment,
+            } => {
+                format!("EV_ {} \"{}\"", env_var_name, comment,)
+            }
+            Self::Plain { comment } => format!("\"{}\"", comment),
+        };
     }
 
     fn parse(s: &str) -> IResult<&str, Self>
-        where
-            Self: Sized {
+    where
+        Self: Sized,
+    {
         let (s, _) = multispace0(s)?;
         let (s, _) = tag("CM_")(s)?;
         let (s, _) = parser::ms1(s)?;
@@ -592,3 +644,52 @@ impl DBCString for Comment {
     }
 }
 
+#[test]
+fn signal_comment_test() {
+    let def1 = "CM_ SG_ 193 KLU_R_X \"This is a signal comment test\";\n";
+    let message_id = MessageId::Standard(193);
+    let comment1 = Comment::Signal {
+        message_id,
+        signal_name: "KLU_R_X".to_string(),
+        comment: "This is a signal comment test".to_string(),
+    };
+    let (_, comment1_def) =
+        Comment::parse(def1).expect("Failed to parse signal comment definition");
+    assert_eq!(comment1, comment1_def);
+}
+
+#[test]
+fn message_definition_comment_test() {
+    let def1 = "CM_ BO_ 34544 \"Some Message comment\";\n";
+    let message_id = MessageId::Standard(34544);
+    let comment1 = Comment::Message {
+        message_id,
+        comment: "Some Message comment".to_string(),
+    };
+    let (_, comment1_def) =
+        Comment::parse(def1).expect("Failed to parse message definition comment definition");
+    assert_eq!(comment1, comment1_def);
+}
+
+#[test]
+fn node_comment_test() {
+    let def1 = "CM_ BU_ network_node \"Some network node comment\";\n";
+    let comment1 = Comment::Node {
+        node_name: "network_node".to_string(),
+        comment: "Some network node comment".to_string(),
+    };
+    let (_, comment1_def) = Comment::parse(def1).expect("Failed to parse node comment definition");
+    assert_eq!(comment1, comment1_def);
+}
+
+#[test]
+fn env_var_comment_test() {
+    let def1 = "CM_ EV_ ENVXYZ \"Some env var name comment\";\n";
+    let comment1 = Comment::EnvVar {
+        env_var_name: "ENVXYZ".to_string(),
+        comment: "Some env var name comment".to_string(),
+    };
+    let (_, comment1_def) =
+        Comment::parse(def1).expect("Failed to parse env var comment definition");
+    assert_eq!(comment1, comment1_def);
+}
