+++ src/message.rs
@@ -6,10 +6,10 @@ extern crate serde_derive;
 
 use derive_getters::Getters;
 
-use crate::DBCString;
 use crate::nodes::Transmitter;
-use crate::signal::Signal;
 use crate::parser;
+use crate::signal::Signal;
+use crate::DBCString;
 
 use nom::{
     bytes::complete::tag,
@@ -43,12 +43,13 @@ impl DBCString for MessageId {
         return match self {
             Self::Standard(id) => id.to_string(),
             Self::Extended(id) => id.to_string(),
-        }
+        };
     }
 
     fn parse(s: &str) -> nom::IResult<&str, Self>
-        where
-            Self: Sized {
+    where
+        Self: Sized,
+    {
         let (s, parsed_value) = complete::u32(s)?;
 
         if parsed_value & (1 << 31) != 0 {
@@ -59,6 +60,32 @@ impl DBCString for MessageId {
     }
 }
 
+#[test]
+fn standard_message_id_test() {
+    let (_, extended_message_id) = MessageId::parse("2").unwrap();
+    assert_eq!(extended_message_id, MessageId::Standard(2));
+}
+
+#[test]
+fn extended_low_message_id_test() {
+    let s = (2u32 | 1 << 31).to_string();
+    let (_, extended_message_id) = MessageId::parse(&s).unwrap();
+    assert_eq!(extended_message_id, MessageId::Extended(2));
+}
+
+#[test]
+fn extended_message_id_test() {
+    let s = (0x1FFFFFFFu32 | 1 << 31).to_string();
+    let (_, extended_message_id) = MessageId::parse(&s).unwrap();
+    assert_eq!(extended_message_id, MessageId::Extended(0x1FFFFFFF));
+}
+
+#[test]
+fn extended_message_id_test_max_29bit() {
+    let s = u32::MAX.to_string();
+    let (_, extended_message_id) = MessageId::parse(&s).unwrap();
+    assert_eq!(extended_message_id, MessageId::Extended(0x1FFFFFFF));
+}
 
 /// CAN message (frame) details including signal details
 #[derive(Clone, Debug, PartialEq, Getters)]
@@ -66,16 +93,17 @@ impl DBCString for MessageId {
 pub struct Message {
     /// CAN id in header of CAN frame.
     /// Must be unique in DBC file.
-    pub (crate) message_id: MessageId,
-    pub (crate) message_name: String,
-    pub (crate) message_size: u64,
-    pub (crate) transmitter: Transmitter,
-    pub (crate)  signals: Vec<Signal>,
+    pub(crate) message_id: MessageId,
+    pub(crate) message_name: String,
+    pub(crate) message_size: u64,
+    pub(crate) transmitter: Transmitter,
+    pub(crate) signals: Vec<Signal>,
 }
 
 impl DBCString for Message {
     fn dbc_string(&self) -> String {
-        return format!("BO_ {} {}: {} {}\n  {}",
+        return format!(
+            "BO_ {} {}: {} {}\n  {}",
             self.message_id.dbc_string(),
             self.message_name,
             self.message_size,
@@ -86,12 +114,13 @@ impl DBCString for Message {
                 .map(|sg| sg.dbc_string())
                 .collect::<Vec<String>>()
                 .join("\n  ")
-        )
+        );
     }
 
     fn parse(s: &str) -> nom::IResult<&str, Self>
-        where
-            Self: Sized {
+    where
+        Self: Sized,
+    {
         let (s, _) = multispace0(s)?;
         let (s, _) = tag("BO_")(s)?;
         let (s, _) = parser::ms1(s)?;
@@ -115,4 +144,11 @@ impl DBCString for Message {
             }),
         ))
     }
+}
+
+#[test]
+fn message_definition_test() {
+    let def = "BO_ 1 MCA_A1: 6 MFA\r\nSG_ ABC_1 : 9|2@1+ (1,0) [0|0] \"x\" XYZ_OUS\r\nSG_ BasL2 : 3|2@0- (1,0) [0|0] \"x\" DFA_FUS\r\n x";
+    Signal::parse("\r\n\r\nSG_ BasL2 : 3|2@0- (1,0) [0|0] \"x\" DFA_FUS\r\n").expect("Failed");
+    let (_, _message_def) = Message::parse(def).expect("Failed to parse message definition");
 }
